# Use `@simd` in `_vecdot`