@KristofferC 2019-04-30
Tangential ref> https://github.com/JuliaArrays/StaticArrays.jl/pull/494.