@tkoolen 2019-04-30
> Tangential ref: #494.

Ah right, I'd forgotten about that PR, it's a little more than tangentially related. Well, this one's up to date with master, and I think the loop range change and getting rid of `@generated` are slight improvements.