@c42f 2019-04-30
Looks nice and clean to me, provided the optimizer is happy with that modification :+1: