@tkoolen 2019-04-30
I'll leave this open for comments for another day and then I'll merge. Thanks for the quick reviews everybody!

After that, perhaps we can try the other reduction operations from #494 again using this approach.