@tkoolen 2019-04-30
I decided to expand the scope of this PR just a little bit to include `bilinear_vecdot`, since it's so similar. There was in fact a lot of code duplication, so I merged `_vecdot` with `_bilinear_vecdot`. I checked that benchmark results for `dot` are still the same.

In doing so, I was wondering whether `(x, y) -> adjoint(x) * y` shouldn't just be replaced with `dot`, given that that's what the generic `dot` implementation in Base does: https://github.com/JuliaLang/julia/blob/b4e1d0eef9c8f2fe3a6e43847bb96cceeac4f4fc/stdlib/LinearAlgebra/src/generic.jl#L758.