@c42f 2019-05-01
CI errors seem related to latest julia, not this PR?