@tkoolen 2019-05-01
Yes, the cron job build of master also fails on Julia nightly.
https://travis-ci.org/JuliaArrays/StaticArrays.jl/builds/526388464